Rename SortDialog to MemoSortPanel
